Re #5 Add Pro Trinket 5V example
This C++11 application is an implementation of the Blink demo for the
Adafruit Pro Trinket 5V.  When run, it causes the LED on pin 13 to blink
at 0.5 Hz.  The mapping of board pins to MCU registers was adapted from
the implementation of the Wiring library for the Arduino Uno.
Compilation and upload flags were inspired by those used by the Arduino
IDE for this board.