Re #7 Add Feather 32u4 example
This C++ application is an implementation of the Blink demo for the
Adafruit Feather 32u4 Basic Proto and other Adafruit ATmega 32u4 boards.
When run, it causes the LED on pin 13 to blink at 0.25 Hz.  Note that
the `PROGRAMMER_PORT` variable in the makefile must be adjusted for a
user's individual environment and may change whenever the board is
plugged in or reset.

The implementation of the `GpioPin` template was inspired by the Wiring
library, and the compilation and upload flags were inspired by those
used by the Arduino IDE for this board.