Re #9 Add Trinket 8 MHz example
This C++ application is an implementation of the Blink demo for the
Adafruit Gemma and Adafruit Trinket running at 8 MHz.  When run, it
causes the LED on pin 1 (D1 on Gemma) to blink at 1 Hz.

As described at
[Programming with AVRdude](https://learn.adafruit.com/introducing-trinket/programming-with-avrdude),
some tweaks to the default avrdude configuration are required when
programming a Trinket.  This example includes an appropriate
"avrdude.conf" for use with Ubuntu Linux, but an alternate file can be
used by changing the `AVRDUDE_CONFIG` variable in the Makefile.

The implementation of the `GpioPin` template was inspired by the Wiring
library, and the compilation and upload flags were inspired by those
used by the Arduino IDE for this board.